Change option_value callbacks from after_save to after_commit
Updating the associated option_values will now happen outside of the initial transaction, which should reduce lock contention by reducing the length of the transaction.

What happens if the original save transaction fails (resulting in a rollback)? In that case, after_commit will not be called and the option_values would ultimately not be updated (which is what would happen anyway if we used after_save and then rolled everything back).

Note: the reason for this change is that we're hitting deadlocks here when updating line items, and the backtrace points to these option_value updates in these callbacks.

Note 2: after_save only triggers on create and update, whereas after_commit also runs on delete, so we specify `on: [:create, :update]` here to keep the behaviour in line with the old callback.