Catch failed transitions after authorization
Both model errors and authorization errors lead to the same user error
message but they aren't the same. If treat them differently as they are
the code is more clear.

It also, lets us choose if we want to show a different error to the user
or even, to show it at all. I'd prefer to show up in the error tracker
in that case.

This patch relies on the following StateMachine's line.

```rb
machine.define_helper(:instance, "#{qualified_name}!") do |machine, object, *args|
  object.send(qualified_name, *args) || raise(StateMachines::InvalidTransition.new(object, machine, name))
end
```
https://github.com/state-machines/state_machines/blob/6a9bb977e0b599eef49fc1fea31b12af113905da/lib/state_machines/event.rb#L225-L227

Failing to transition to pending will raise an exception we can catch.
No need for extra error handling logic.