Remove accidental modification of the db schema
fa74dae99c77f1b764a164241970637b5d82cbbb changed the schema file even
though the migration doesn't touch the database structure. A previous
development version of the migration did this but the finally merged
version didn't. So this schema is not valid. Let's hope that nobody set
up a new instance with the changed schema and we just fix it to what it
was before.