Add entity name to Enterprises
For the Brazilian context, it's good to know if an enterprise is
associated with an entity.

For simplicity, we'll start by adding only a field with the entity name.

Later, we'll add it to the enterprise admin page so it'll be possible to
edit it. We'll build a list that will be loaded on the view because
there isn't enough time to implement an amin page to manage entities.