Extract entity options to a helper
The logic is better placed there. This way, if we change the way we
build the list, say by finding names in the database, we have a place to
change the logic (and possibly reuse it at other places).

Also it makes it easier to later unit test the method so we have
guarantees it works.