Bump Node to v14.21.2
Currently our CI and servers do not respect this setting, only dev environments that have a Node version manager do.

I checked a few prod servers, and they use this version already so now I'm making it official. But we should aim to upgrade further, the current version is v19.

    WARNING: node-v14.21.2-darwin-x64 is in LTS Maintenance mode and nearing its end of life.
    It only receives *critical* security updates, *critical* bug fixes and documentation updates.