Remove unsupported config
In Sidekiq v7, this field is not recognised and gets passed through to Redis, which also doesn't recognise it and crashes.

It was added to avoid filling up Redis on dev, but it's actually a Sidekiq pro option which needs to be defined on jobs, so I guess it never worked.