[wip] Use TableComponent to show products
Ok, how do we translate product objects into a table? The component just wants a 2D array. But we want to do clever things using an ActiveModel. So should we pre-generate the 2D array? How will we handle form elements in the future? Probably row-by row. So there needs to be a row component. And a BulkEditProductRow could inherit from it and implement the necessary elements. So we build a collection of product rows and pass it in?

Let's see how Polaris does it.