Try out Polaris data table
The newer version of Polaris view components was needed. Dunno why bundler didn't install it to begin with.

But, it's unstyled. Can we import just the related styles from Polaris? No, polaris_view_components is a port of Polaris (which is built in React), and actually contains a generated export of the full polaris stylesheet (https://github.com/baoagency/polaris_view_components/blob/main/app/assets/stylesheets/polaris_view_components.css)

So the styles aren't componentised and we can't separate them out. It looks like the Javascript [is](https://github.com/baoagency/polaris_view_components/blob/main/app/assets/javascripts/polaris_view_components/data_table_controller.js), but it doesn't seem to have any features of any consequence. I don't see how the sorting works, and anyway we need to fetch sorted data from the server, which it may not support out of the box.

We can still benefit from the well-structured HTML. It includes a lot of Polaris classnames, and it could be beneficial to commit to this design system and extract more useful bits in the future.
But it's also a third-party system which we will end up having to extend pretty heavily for our own purposes. So I don't predict there'll be much benefit from using it as a foundation.