Fix payment_required? when a voucher is applied to an order
In the case when an order total is 0 due to a voucher being applied, and
going through a checkout after a cart update, the order was confirmed
autmatically and the customer was redirected to the homepage.
In this scenario the order was considered a "Free order". To fix the issue
we require payment for any order with a voucher applied.