Clip long cell content
Columns are arranged with percentage, so that they will fill to the full available width, regardless of screen size.
Some columns have max widths because their content will never be bigger (eg dates and IDs).

Another option was to use line-clamp inside the cells, but table-layout:fixed is more efficient. Oh, I think we'll need that if columns are dynamtically shown/hidden.