Rename color-1 to white
I considered doing this to all files, but decided it will be better to avoid touching files outside admin_v3.
So I considered doing this in master first. But then we have do do it here again anyway.
So I just did it this way. Time to just get stuff done.