Fix require order on dfc_provider engine
Previously the test expected a class to already be defined in order to
load swagger test config, without explicitly loading the
base_spec_helper (that loads the app, and the classes).

Occasionally, depending on the order of execution, the app could be loaded
if another test required the spec_helper / base_spec_helper.