Only count changed products
I'd much rather update the frontend to avoid submitting unchanged fields in the first place.

So I tried to do that with the current StimulusReflex form submit logic, but was unable to get it to work (documentation seems to be out of date). I think we shouldn't be submitting the form with StimulusReflex anyway.

So I tried changing it to a standard form submit. But was unable to get it to work. So.. we have this ugly code instead sorry.