Add SemanticLink model for variants
We want to link variants/products to external DFC SuppliedProducts to
trigger supplier orders when local stock is exhausted. This is the first
step to enable the link.

Since we want to order wholesale variants of the product, we need a
quantity field as well. This is only temporary until we receive the
quantity through `hasIngredient` on the DFC API.